feat: :zap: Restructuring the create albums form
Created album creation and editing form component on the playlist screen, to reduce code

Improvement will be added to receive data and send data
