+++ front/src/components/Forms/FormPlaylist/index.tsx
@@ -1,0 +1,75 @@
+import Modal from "@/components/Dialogues/Modal";
+import Input from "@/components/Input";
+import Select from "@/components/Select";
+import Button from "@/components/Button";
+import { useState } from "react";
+import { useFormState } from "react-dom";
+import { AlbumAction } from '@/data/actions/albums-actions/playlist-actions';
+import { fetchWrapper } from '../../../app/functions/fetch'
+
+interface FormPlaylistProps {
+    editAlbum: boolean
+    openCard: boolean;
+}
+
+const INITIAL_STATE = {
+    data: null,
+    zodErrors: null,
+    message: null,
+};
+
+const FormPlaylist:React.FC<FormPlaylistProps> = ({ editAlbum, openCard }) => {
+    
+    const [artist, setArtist] = useState<any[]>([])
+    const [nameAlbum, setNameAlbum] = useState("")
+    const [yearAlbum, setYearAlbum] = useState("")
+
+    const [formState, formAction] = useFormState(
+        AlbumAction,
+        INITIAL_STATE
+    );
+    
+    return(
+        <>
+            <Modal
+                openModal={openCard}
+            >
+                <div className="mb-10 sm:mx-auto sm:w-full sm:max-w-sm">
+                    <form 
+                        action={formAction}
+                        className='space-y-10'
+                    >
+                        <Input
+                            required={true}
+                            label="Name of Album"
+                            placeholder="Enter name of your album: "
+                            id="nameAlbum"
+                            name="nameAlbum"
+                            type="nameAlbum"
+                            error={formState?.zodErrors?.nameAlbum}
+
+                        />
+                        <Select
+                            label="Artist of Album"
+                            items={artist}         
+                        />
+                        <Input
+                            required={true}
+                            label="Year of Album"
+                            placeholder="Enter year of your album: "
+                            id="yearAlbum"
+                            name="yearAlbum"
+                            type="yearAlbum"
+                            error={formState?.zodErrors?.yearAlbum}
+                        />
+                        <Button
+                            text={editAlbum ? 'Edit this album' : 'Save new album'}
+                        />
+                    </form>
+                </div>
+            </Modal>
+        </>
+    )
+}
+
+export default FormPlaylist;
